JSON file of patient IDs and images
This is a JSON file in which each patient ID corresponds to a dictionary, the keys of which are the four image modalities (FLAIR, T1w, T1wCE, T2w), and the values are lists of the image file names that are not blank & in physical order according to the "Slice Location" data in the corresponding DICOM file (access via pydicom.dcmread(img)[('0020','1041')].value).  
Not all patients are included, there are 14 in which the DICOM files for at least one image modality did not have the requisite information to put the images in physical order.  Those haven't been dealt with yet.  There is also at least one patient, ID 00709, whose FLAIR image list is empty in this JSON file.  This may be because those images are all blank, but for the time being it has been added to the "weirds" list along with the other 14 which are excluded from training for now.