Helper function to create the patient JSON file
This is a function-version of the code used to create the JSON file with the patient images in physical order.  The "excludes" had to be found experimentally by attempting to run the main for-loop in this function many times, noting which patient caused the crash & adding them to the exclude  (or "weird") list individually.  The line that would cause the crash is:
image_slice_locs = [pydicom.dcmread(im)[('0020', '1041')].value for im in imgnames]
It would crash because some of the DICOM files do not have the ('0020','1041') ('Slice Location') data.  Realistically there is away to have this function automatically skip those patients, add their names to a "weirds" list, and continue on all in one shot, but as there are less than 600 patients, the weird ones are rare, and I only need to complete this process **once** it didn't seem worth it to do that.