Helper function DICOM -> np array
This function takes in the path to a DICOM file and produces a normalized np array image output.  It also applies the VOI_LUT transform if desired, and fixes an issue with the MONOCHROME1 mode if present & the fix is desired.  
This function WAS NOT WRITTEN BY ME,  credit goes to: https://www.kaggle.com/furcifer/no-baseline-pytorch-cnn-for-mri