Notebook as of September 1st
This one has more EDA in it, as well as a full training loop and MultiSliceLoader, which is a generalization of the existing SliceLoader for selecting more than one image per patient.  At the moment MultiSliceLoader takes several consecutive images from the middle of the set of images for the given patient & modality, but this will be upgraded to target particular areas in future versions.