Added feature finder, multi-mode model
Added a new submodel, FeatureFinder, which takes in the 3D image for a particular mode, finds some low-level features, and then feeds those to a bigger model to be combined with the low-level features from the other modalities.  I have pretty well determined that the only hope is to isolate the tumor slices, so that's what will be done next.